+++ packages/opentelemetry-exporter-collector/src/platform/browser/types.ts
@@ -1,0 +1,25 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { CollectorExporterConfigBase } from '../../types';
+
+/**
+ * Collector Exporter Config for Web
+ */
+export interface CollectorExporterConfigBrowser
+  extends CollectorExporterConfigBase {
+  headers?: { [key: string]: string };
+}
