+++ packages/opentelemetry-exporter-collector/src/platform/node/types.ts
@@ -16,7 +16,10 @@
 
 import * as grpc from 'grpc';
 import { ReadableSpan } from '@opentelemetry/tracing';
-import { CollectorExporterError } from '../../types';
+import {
+  CollectorExporterError,
+  CollectorExporterConfigBase,
+} from '../../types';
 
 /**
  * Queue item to be used to save temporary spans in case the GRPC service
@@ -37,4 +40,13 @@ export interface ServiceClient extends grpc.Client {
     metadata: grpc.Metadata | undefined,
     callback: Function
   ) => {};
+}
+
+/**
+ * Collector Exporter Config for Node
+ */
+export interface CollectorExporterConfigNode
+  extends CollectorExporterConfigBase {
+  credentials?: grpc.ChannelCredentials;
+  metadata?: grpc.Metadata;
 }
